+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `split` function that splits a string by a separator into an array of strings.
+- `toLower` function that converts a string to lowercase.
+- `toUpper` function that converts a string to uppercase.
 
 ### Fixed